Unify config TOML serialization
