+++ tests/config_to_toml.rs
@@ -115,7 +115,7 @@ fn to_toml_includes_fix_policy() {
     let cfg_path = td.path().join(".ryl.toml");
     fs::write(
         &cfg_path,
-        "[fix]\nfixable = ['ALL', 'braces', 'brackets', 'commas', 'comments', 'comments-indentation', 'new-lines']\nunfixable = ['braces', 'new-line-at-end-of-file']\n",
+        "[fix]\nfixable = ['ALL', 'braces', 'brackets', 'commas', 'comments', 'comments-indentation', 'new-line-at-end-of-file', 'new-lines']\nunfixable = ['braces', 'brackets', 'commas', 'comments', 'comments-indentation', 'new-line-at-end-of-file', 'new-lines']\n",
     )
     .unwrap();
 
@@ -129,6 +129,7 @@ fn to_toml_includes_fix_policy() {
     .unwrap();
 
     let toml = ctx.config.to_toml_string().unwrap();
+    assert!(toml.contains("yaml-files = ["));
     assert!(toml.contains("[fix]"));
     assert!(toml.contains("fixable = ["));
     assert!(toml.contains("\"ALL\""));
@@ -137,8 +138,14 @@ fn to_toml_includes_fix_policy() {
     assert!(toml.contains("\"commas\""));
     assert!(toml.contains("\"comments\""));
     assert!(toml.contains("\"comments-indentation\""));
+    assert!(toml.contains("\"new-line-at-end-of-file\""));
     assert!(toml.contains("\"new-lines\""));
     assert!(toml.contains("unfixable = ["));
     assert!(toml.contains("\"braces\""));
+    assert!(toml.contains("\"brackets\""));
+    assert!(toml.contains("\"commas\""));
+    assert!(toml.contains("\"comments\""));
+    assert!(toml.contains("\"comments-indentation\""));
     assert!(toml.contains("\"new-line-at-end-of-file\""));
+    assert!(toml.contains("\"new-lines\""));
 }
